Move setup_runtimelink_test test method into its only caller. NFC.
This code is pretty specific and has just one caller. I can't
imagine other tests wanting to share this exact code.
